% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRDemographic.R
\name{wpCPRDemographic}
\alias{wpCPRDemographic}
\title{wpCPRDemographic function to get a toptal subnational population age-sex structures
based on shapefile}
\usage{
wpCPRDemographic(
  year = 2000,
  shapeFilePath = NULL,
  outputFilePath = NULL,
  apikey = NULL,
  callbacktime = 5,
  maxexectime = 3600,
  apiurl = NULL,
  verbose = FALSE
)
}
\arguments{
\item{year}{The year of the population (2000-2020)}

\item{shapeFilePath}{The path to the shpaefile}

\item{outputFilePath}{The opath to the output CVS file}

\item{apikey}{API key to access the WorldPop API}

\item{callbacktime}{Default is 5 sec. TIme to call the API server}

\item{maxexectime}{Default is 3600 sec. Max execution time of the request.}

\item{apiurl}{URL to WorldPop API}

\item{verbose}{If TRUE then the progress will be shown}
}
\value{
DataFrame* object
}
\description{
wpCPRDemographic function to get a toptal subnational population age-sex structures
based on shapefile
}
\examples{
wpCPRDemographic(year=2013, shapeFilePath = file.path("/tmp/Inputs/BFA-HDX-Borders.shp"))
	
}
