% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRGetBaseURL.R
\name{wpCPRGetBaseURL}
\alias{wpCPRGetBaseURL}
\title{wpCPRGetBaseURL function to return a URL of API}
\usage{
wpCPRGetBaseURL(url)
}
\arguments{
\item{url}{URL path to WorldPop API}
}
\value{
string
}
\description{
wpCPRGetBaseURL function to return a URL of API
}
