% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRGetlog.R
\name{wpCPRGetlog}
\alias{wpCPRGetlog}
\title{wpCPRGetlog function will return a data.frame}
\usage{
wpCPRGetlog(file = NULL)
}
\arguments{
\item{file}{The path to the shpaefile}
}
\value{
A data.frame with a log data of last API calls
}
\description{
wpCPRGetlog function will return a data.frame
}
\examples{
wpCPRGetlog()

}
