% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRGetTask.R
\name{wpCPRGetTask}
\alias{wpCPRGetTask}
\title{wpCPRGetTask function to get respond}
\usage{
wpCPRGetTask(path = "v1/tasks/", taskid)
}
\arguments{
\item{path}{Sub-path to the API}

\item{taskid}{Task ID of the job on WorldPop API}
}
\value{
json
}
\description{
wpCPRGetTask function to get respond
}
