% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRPostRequest.R
\name{wpCPRPostRequest}
\alias{wpCPRPostRequest}
\title{wpCPRPostRequest function to send POST request}
\usage{
wpCPRPostRequest(path = "v1/services/stats", args)
}
\arguments{
\item{path}{Sub-path to the API}

\item{args}{Extra arguments for httr::POST}
}
\value{
string
}
\description{
wpCPRPostRequest function to send POST request
}
