% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRProgressMessage.R
\name{wpCPRProgressMessage}
\alias{wpCPRProgressMessage}
\title{wpCPRProgressMessage function will return progress}
\usage{
wpCPRProgressMessage(x, max = 100, label = NULL)
}
\arguments{
\item{x}{integer}

\item{max}{maximum for progress bvar}

\item{label}{additional text for progress  bar}
}
\value{
character
}
\description{
Tested on Windows 10
}
\examples{
wpCPRProgressMessage( x=10, max = 200, label="Progress message" )
}
