% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPRTimeDiff.R
\name{wpCPRTimeDiff}
\alias{wpCPRTimeDiff}
\title{wpCPRTimeDiff function to get time difference in human readable format
Input is start time and end time
If "frm" is set to "hms" then output will be h:m:s
otherwise only hours will be returned}
\usage{
wpCPRTimeDiff(start, end, frm = "hms")
}
\arguments{
\item{start}{Starting time}

\item{end}{Ending time}

\item{frm}{If "frm" is set to "hms" then output will be h:m:s}
}
\value{
numeric
}
\description{
wpCPRTimeDiff function to get time difference in human readable format
Input is start time and end time
If "frm" is set to "hms" then output will be h:m:s
otherwise only hours will be returned
}
