% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpTimeDiff.R
\name{wpTimeDiff}
\alias{wpTimeDiff}
\title{wpTimeDiff function to get time difference in human readable format
Input is start time and end time
If "frm" is set to "hms" then output will be h:m:s
otherwise only hours will be returned}
\usage{
wpTimeDiff(start, end, frm = "hms")
}
\value{
numeric
}
\description{
wpTimeDiff function to get time difference in human readable format
Input is start time and end time
If "frm" is set to "hms" then output will be h:m:s
otherwise only hours will be returned
}
