% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpCPR.R
\name{wpCPRPopulation}
\alias{wpCPRPopulation}
\title{wpCPRPopulation function to get a toptal population
based on shapefile}
\usage{
wpCPRPopulation(
  year = 2000,
  shapefile = NULL,
  layer = NULL,
  attribute_key = NULL,
  outputCVSDir = NULL,
  addpopshp = FALSE,
  api_key = NULL,
  callback_time = 5,
  max_exec_time = 600,
  wp_API_url = NULL,
  verbose = FALSE
)
}
\arguments{
\item{year}{The year of the population (2000-2020)}

\item{shapefile}{The path to the shpaefile}

\item{layer}{The layer name.}

\item{attribute_key}{The name of the attribute in the SHP file}

\item{outputCVSDir}{The opath to the output CVS file}

\item{addpopshp}{If TRUE then the total population of each polygon will be added to a new SHP file}

\item{api_key}{API key to access the WorldPop API}

\item{callback_time}{Default is 5 sec. TIme to call the API server}

\item{max_exec_time}{Default is 600 sec. Max execution time of the request.}

\item{wp_API_url}{URL to WorldPop API}

\item{verbose}{If TRUE then the progress will be shown}
}
\value{
DataFrame* object
}
\description{
wpCPRPopulation function to get a toptal population
based on shapefile
}
\examples{
wpCPRPopulation(year=2013, shapefile = file.path("/tmp/Inputs/BFA-HDX-Borders.shp"),attribute_key="admin2Pcod")
	
}
